refactor: extract runtime artifact specs
